third_party/blink/web_tests/fast/forms/setCustomValidity-expected.txt

This test checks if setCustomValidity causes CSS to get applied correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

  
PASS test.validity.customError is false
PASS test.validationMessage is ""
PASS getComputedStyle(test).backgroundColor is "rgb(0, 255, 0)"
PASS test.validity.customError is true
PASS test.validationMessage is "Custom validation message"
PASS getComputedStyle(test).backgroundColor is "rgb(255, 0, 0)"
PASS test.validity.customError is false
PASS test.validationMessage is ""
PASS getComputedStyle(test).backgroundColor is "rgb(0, 255, 0)"
PASS test.validity.customError is false
PASS test.validationMessage is ""
PASS getComputedStyle(test).backgroundColor is "rgb(0, 255, 0)"
PASS test.validity.customError is true
PASS test.validationMessage is "Custom validation message"
PASS getComputedStyle(test).backgroundColor is "rgb(255, 0, 0)"
PASS test.validity.customError is false
PASS test.validationMessage is ""
PASS getComputedStyle(test).backgroundColor is "rgb(0, 255, 0)"
PASS test.validity.customError is false
PASS test.validationMessage is ""
PASS getComputedStyle(test).backgroundColor is "rgb(0, 255, 0)"
PASS test.validity.customError is true
PASS test.validationMessage is "Custom validation message"
PASS getComputedStyle(test).backgroundColor is "rgb(255, 0, 0)"
PASS test.validity.customError is false
PASS test.validationMessage is ""
PASS getComputedStyle(test).backgroundColor is "rgb(0, 255, 0)"
PASS successfullyParsed is true

TEST COMPLETE
